refactor usePresetActionHandler in a separate hook
